+++ typescript/agentkit/src/wallet-providers/index.ts
@@ -5,3 +5,5 @@ export * from "./cdpWalletProvider";
 export * from "./svmWalletProvider";
 export * from "./solanaKeypairWalletProvider";
 export * from "./privyWalletProvider";
+export * from "./privyEvmWalletProvider";
+export * from "./privySvmWalletProvider";